Gemini-agentic-search/requirements.txt
google-genai==0.5.0
python-decouple==3.8
pydantic==2.12.4
